requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
cycler==0.11.0
fonttools==4.29.1
h11==0.13.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.5.5.62
outcome==1.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
wsproto==1.0.0
